+++ Project1_Starter_Code/Rule.cpp
@@ -4,7 +4,7 @@
 
 Rule::Rule(Predicate copyHeadPredicate, std::vector<Predicate*> copyPredicates){
     headPredicate = copyHeadPredicate;
-    for (int i = 0; i < copyPredicates.size(); ++i){
+    for (int i = 0; i < (int)copyPredicates.size(); ++i){
         bodyPredicates.push_back(copyPredicates.at(i));
     }
 }
@@ -12,8 +12,8 @@ Rule::Rule(Predicate copyHeadPredicate, std::vector<Predicate*> copyPredicates){
 std::string Rule::ToString() {
     std::stringstream returnString;
     returnString << headPredicate.ToString(headPredicate) << " :- ";
-    for (int i = 0; i < bodyPredicates.size(); i++){
-        if (i != bodyPredicates.size() -1) {
+    for (int i = 0; i < (int)bodyPredicates.size(); i++){
+        if (i != (int)bodyPredicates.size() -1) {
             returnString << bodyPredicates.at(i)->ToString(*bodyPredicates.at(i)) << ",";
         }
         else {
